Dealership can sell car to customer.

+++ src/main/java/vehicles/Car.java
@@ -4,7 +4,7 @@ public class Car extends Vehicle {
 
     private int engineSize;
 
-    public Car(String color, double price, BodyType bodyType, int engineSize) {
+    public Car(String color, int price, BodyType bodyType, int engineSize) {
         super(color, price, bodyType);
         this.engineSize = engineSize;
     }
