+++ src/main/java/vehicles/ElectricCar.java
@@ -4,7 +4,7 @@ public class ElectricCar extends Vehicle{
 
     private int chargingTime;
 
-    public ElectricCar(String color, double price, BodyType bodyType, int chargingTime) {
+    public ElectricCar(String color, int price, BodyType bodyType, int chargingTime) {
         super(color, price, bodyType);
         this.chargingTime = chargingTime;
     }
