+++ src/main/java/vehicles/HybridCar.java
@@ -4,7 +4,7 @@ public class HybridCar extends Vehicle {
 
     private int driveTimeOnBattery;
 
-    public HybridCar(String color, double price, BodyType bodyType, int driveTimeOnBattery) {
+    public HybridCar(String color, int price, BodyType bodyType, int driveTimeOnBattery) {
         super(color, price, bodyType);
         this.driveTimeOnBattery = driveTimeOnBattery;
     }
